setup flask server
