--- visualization/package.json
@@ -1,22 +1,0 @@
-{
-    "name": "heatmap-visualization",
-    "version": "1.0.0",
-    "description": "",
-    "main": "src/index.js",
-    "scripts": {
-        "build": "rm -rf dist && webpack",
-        "dev": "webpack-dev-server",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "author": "",
-    "license": "ISC",
-    "devDependencies": {
-        "html-webpack-plugin": "^3.2.0",
-        "webpack": "^4.42.0",
-        "webpack-cli": "^3.3.11",
-        "webpack-dev-server": "^3.10.3"
-    },
-    "dependencies": {
-        "d3": "^5.15.0"
-    }
-}
