--- visualization/webpack.config.js
@@ -1,16 +1,0 @@
-const path = require("path");
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-
-module.exports = {
-  entry: "./src/index.js",
-  output: {
-    path: path.resolve(__dirname, "dist"),
-    filename: "bundle.[hash].js"
-  },
-  devtool: "inline-source-map",
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: path.resolve(__dirname, "template.html")
-    })
-  ]
-};
